refactor: converts Storybook preview.jsx to typescirpt, extracts decorator
